Clean up SOP derivations, create designated modules
